// src/app/moodle/apis/candidate.ts
export interface Candidate {
    id?: number;
    username?: string;
    password?: string;
    firstname?: string;
    middlename?: string;
    lastname?: string;
    alternatename?: string;
    email?: string;
    age?: number;
    academicdegreeid?: number;
    positionid?: number;
    status?: boolean;
    name?: string;
}